+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## 6.4.19
-
+### Enhancements
+- Move to `sdk-go-bundle` for logging sdk
 ### Fixes 
 - Fail on k8s cluster and nodepool if creation or deletion entered failed state
 - K8s and dataplatform tests
 ## 6.4.18
-
 ### Features
 - Add tests for Mongo cluster and user
 - Add new fields for NICs in VM Autoscaling group (firewall_active, firewall_type, firewall_rule, flow_log, target_group)